+++ pkg/coredata/migrations/20250417T085900Z.sql
@@ -1,0 +1,7 @@
+ALTER TABLE evidences DROP CONSTRAINT evidences_task_id_fkey;
+
+ALTER TABLE evidences
+  ADD CONSTRAINT evidences_task_id_fkey
+  FOREIGN KEY (task_id)
+  REFERENCES tasks(id)
+  ON DELETE CASCADE;